Refactor __init__ method of metrics serializers to fix bugs
This commit addresses several bugs that were caused by the
__init__ methods of the metrics serializers. The previous
implementation of this method had a vulnerability that allowed
users, under specific conditions, to retrieve metrics for all
users in the database.

The changes on this commit fix the issue by only using the
queryset used to initialized the serialzer and by removing the
default queryset with the full set of trades.

The bug was being triggered by requesting the metrics for a user
with no registered trades.