Add `countries` with endpoints for countries, indicators and reports ...
The new `countries` app is intended to be used as a way to keep a record
of the economic indicators of each nation/country of interest and, to
register events/reports on which new data, with regard to the indicators,
will be released.

The countries app, provides several endpoints that allow to create, read,
and edit, objects for countries, economic indicators, and economic reports.